#103 refactor: StoolLogCollectionView의 헤더 높이를 HeaderType을 통한 absolute 값이 아닌, estimated 값으로 설정하도록 변경
